trying to redefine a TransitionAction on a Transition is an error
Overwriting transition actions has caused several problems in the past
few days. We made the decision that a transition can have 0 or 1 actions
associated with it, but no check for blindly setting an existing action
to a new value. If a legitimate need to do such ever arises, please add
a function like transition.replace_action(new_action).